Test Cases for sessions and test scripts
