+++ backend/src/config/database.js
@@ -10,11 +10,3 @@ export function connectDatabase() {
     }).promise()
 }
 
-// export const pool = mysql.createPool({
-//     host:process.env.MYSQL_HOST,
-//     port: process.env.MYSQL_PORT,
-//     user: process.env.MYSQL_USER,
-//     password: process.env.MYSQL_PASSWORD,
-//     database: process.env.MYSQL_DATABASE
-// }).promise();
-
